Fixed packet handling in RedMarker.
RedMarker has to remove Ethernet header and recompute FCS after marking
the packet.

Very dirty approach because it uses static FCS_mode no matter what
FCS_mode is set on the actual interface and assumes Ethernet encap.

+++ src/inet/common/queue/RedDropper.cc
@@ -38,7 +38,7 @@ void RedDropper::initialize()
 {
     AlgorithmicDropperBase::initialize();
 
-    wq = par("wq");
+    wq = par("wq").doubleValue();
     if (wq < 0.0 || wq > 1.0)
         throw cRuntimeError("Invalid value for wq parameter: %g", wq);
